migrate: re-run post-step callbacks on error
This commit modifies the migration framework to re-attempt post-step
callbacks if they error during a migration. Previously, if a post-step
callbacks failed, but their associated SQL migration succeeded, the
post-step callback would not be re-attempted on the next migration run,
and instead proceed with the next SQL migration.

This is achieved by introducing the concept of a "post-step callback"
migration version. Post-step callbacks are their corresponding SQL
migration version offset by +1000000000.

During the execution of a post-step callback, the post-step callback
migration version will be persisted as the database version.
That way, if the post-step callback errors, the version for the database
will be the post-step callback version on the next startup. The
post-step callback will then be re-attempted before proceeding with
the next SQL migration.