migrate: re-run migration tasks callbacks on error
This commit modifies the migration framework to re-attempt migration
tasks if they error during a migration, on the next run. Previously, if
a migration task failed, but their associated SQL migration succeeded,
the database version would be set to a dirty state, and require manual
intervention in order to reset the SQL migration and re-attempt it + the
migration task.

The new re-attempt mechanism is achieved by introducing the concept of a
"migration task" version. Migration task versions are their
corresponding SQL migration version offset by +1000000000.

During the execution of a migration task, the migration task version
will be persisted as the database's version.
That way, if the migration task errors, the version for the database
will be the migration task version on the next startup. The migration
task will then be re-attempted before proceeding with the next SQL
migration.