migrate: re-run migration tasks callbacks on error
This commit modifies the migration framework to re-attempt migration
tasks if they error during a migration, on the next run. Previously, if
a migration task failed, but their associated SQL migration succeeded,
the database version would be set to a dirty state, and require manual
intervention in order to reset the SQL migration and re-attempt it + the
migration task.

The new re-attempt mechanism is achieved by ensuring that a migration
can only be either an SQL migration or a migration task, but not both.
This way, if a migration task errors, the database version will be
reset to the previous version prior to executing the migration task, and
the migration task will be re-attempted on the next run.